Add support for classes with type parameters - WIP
Added tests to verify support for generic parameters. Thing is, when we use a generic type like A we also need to define at least an implicit Typeable for that parameter, or it will fail on compilation, due to diverging implicit resolution. Which turns these onto normal 'defined' parameters.

I push on top of previous PR as to benefit from changes in that PR regarding underlying implementation