Add support for self-recursive type (Fix #5)
The issue seems to be specific to the aim of this library, as we use Typeable at some point to create a Ref() element, and that breaks for recursive case classes like `final case class Recursive(head: Int, tail: Option[Recursive])``

I couldn't find a way to make Typeable accept these recursive types, so I had to leave Shapeless and work with `WeakTypeTag` from scala.runtime.universe. It's hackish, but it works, and it is used only 'at the bundaries' so... I guess it's ok.

Also, this may slow down compilation a bit... sorry!